// client-monorepo/libs/applets/wealth/src/lib/features/funds/models/fund-profile.model.ts
export interface IFundProfileModel {
  dailyEfficiency?: number;
  weekleyEfficiency?: number;
  compoundProfit?: number;
  monthlyEfficiency?: number;
  quarterlyEfficiency?: number;
  sixMonthEfficiency?: number;
  annualEfficiency?: number;
  netAsset?: number;
  websiteAddress?: string;
  statisticalNav?: number;
  issueNav?: number;
  cancelNav?: number;
  latestUpdate?: string;
  title?: string;
  saleNav?: number;
  contractLink?: string;
  dividendPeriod?: string;
  expectedProfit?: string;
  investmentType?: 'Mutual' | 'ETF' | 'CrowdFund';
  latestNavUpdate?: string;
  liquidityPolicy?: string;
  purchaseNav?: number;
  riskLevel?: string;
  statisticNav?: number;
  symbol?: string;
  thumbnailLogoAddress?: string;
  type?: string;
  dividendDescription?: string;
  website?: string;
  maxBuyableAmount?: number;
  minBuyableAmount?: number;
  displaySymbol?: string;
  efficiency?: string;
  status?: string;
  financialReportWebsite?: string;
  industryName?: string;
  companyName?: string;
  oldName?: string;
  buyable: boolean;
  sellable: boolean;
}